+++ src/main/java/racingcar/domain/Cars.java
@@ -8,7 +8,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class Cars implements Iterable<Car> {
+public class  Cars implements Iterable<Car> {
     private final List<Car> cars;
 
     public Cars(List<Car> cars) {
