+++ src/main/java/racingcar/web/controller/RacingCarController.java
@@ -1,10 +1,13 @@
-package racingcar;
+package racingcar.web.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
+import racingcar.web.dto.ResultDto;
+import racingcar.web.dto.UserInputDto;
+import racingcar.web.service.RacingGameService;
 
 @Controller
 public class RacingCarController {
