+++ src/main/java/racingcar/web/dao/GameResultDao.java
@@ -1,30 +1,28 @@
-package racingcar.dao;
+package racingcar.web.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
-import racingcar.entity.CarEntity;
+import racingcar.web.entity.GameResultEntity;
 
 import java.sql.PreparedStatement;
 
 @Component
-public class CarDao {
+public class GameResultDao {
+
     private final JdbcTemplate jdbcTemplate;
 
-    public CarDao(JdbcTemplate jdbcTemplate) {
+    public GameResultDao(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Long save(final CarEntity carEntity){
-        String sql = "insert into car (player_name, final_position, is_winner, game_result_id) values (?, ?, ?, ?)";
+    public Long insert(GameResultEntity gameResultEntity) {
+        String sql = "insert into game_result (try_count) values (?)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
         jdbcTemplate.update(connection -> {
             PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
-            ps.setString(1, carEntity.getPlayerName());
-            ps.setInt(2, carEntity.getFinalPosition());
-            ps.setBoolean(3, carEntity.isWinner());
-            ps.setLong(4, carEntity.getGameResultId());
+            ps.setInt(1, gameResultEntity.getTryCount());
             return ps;
         }, keyHolder);
