+++ src/main/java/racingcar/web/dto/CarDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.web.dto;
 
 public class CarDto {
     private final String name;
