+++ src/main/java/racingcar/web/dto/ResultDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.web.dto;
 
 import racingcar.domain.Cars;
