+++ src/main/java/racingcar/web/dto/UserInputDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.web.dto;
 
 public class UserInputDto {
