+++ src/test/java/racingcar/dao/CarDaoTest.java
@@ -6,7 +6,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import racingcar.domain.TryCount;
-import racingcar.entity.CarEntity;
+import racingcar.web.entity.CarEntity;
+import racingcar.web.dao.CarDao;
+import racingcar.web.dao.GameResultDao;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 class CarDaoTest {
